Fix and Improve permission

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Admin\BaseController;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\DB;
+
 
 class PermissionController extends BaseController
 {
@@ -26,48 +28,37 @@ class PermissionController extends BaseController
 
     public function create()
     {
+        $this->response['guards'] = config('auth.guards') ? array_keys(config('auth.guards')) : ['web'];
         return view('admin.role-permission.permission.create', $this->response);
     }
 
     public function store(Request $request)
     {
         $request->validate([
-            'name' => [
-                'required',
-                'string',
-                'unique:permissions,name'
-            ],
-            'guard_name' => [
-                'required',
-                'string'            ]
+            'name' => 'required|string|unique:permissions,name|min:4|max:16',
+            'guard_name' => 'required|string'
         ]);
 
-        Permission::create([
+        $permission = Permission::create([
             'name' => $request->name,
             'guard_name' => $request->guard_name,
         ]);
 
-        return redirect(route('admin.permission.index'))->with('status','Permission Created Successfully');
+        return redirect(route('admin.permission.edit', [$permission->id]))->with('status','Permission Created Successfully');
     }
 
     public function edit(Permission $permission)
     {
         $this->response['permission'] = $permission;
+        $this->response['guards'] = config('auth.guards') ? array_keys(config('auth.guards')) : ['web'];
         return view('admin.role-permission.permission.update', $this->response);
     }
 
     public function update(Request $request, Permission $permission)
     {
         $request->validate([
-            'name' => [
-                'required',
-                'string',
-                'unique:permissions,name,'.$permission->id
-            ],
-            'guard_name' => [
-                'required',
-                'string'
-            ]
+            'name' => 'required|string|unique:permissions,name,' . $permission->id . '|min:4|max:16',
+            'guard_name' => 'required|string'
         ]);
 
         $permission->update([
@@ -80,8 +71,15 @@ class PermissionController extends BaseController
 
     public function destroy($permissionId)
     {
-        $permission = Permission::find($permissionId);
-        $permission->delete();
-        return redirect(route('admin.permission.index'))->with('status','Permission Deleted Successfully');
+        try {
+            $permission = Permission::findOrFail($permissionId);
+            if ($permission) {
+                $permission->delete();
+                return redirect(route('admin.permission.index'))->with('status','Permission Deleted Successfully');        
+            }
+        } catch (\Exception $e) {
+            dd($e);
+            return redirect(route('admin.permission.index'))->with('status','Permission Couldn\'t be deleted Successfully');
+        }
     }
 }
